Move SD search API to separate package
